create_report/requirements.txt
google-cloud-storage
matplotlib
numpy
pandas
seaborn